requirements.txt
polygon-api-client
Django>=3.0,<4.0
django_table_sort
django_mysql
pandas
requests
datetime
python-dotenv
django-mathfilters
djangorestframework
markdown   
django-filter
django-cors-headers
virtualenv
